requirements.txt
asgiref==3.6.0
dj-rest-auth==4.0.0
Django==4.2.1
django-autocomplete-light==3.9.4
django-cors-headers==3.14.0
django-filter==22.1
django-location-field==2.7.0
django-treebeard==4.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Markdown==3.4.1
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
rest-framework-simplejwt==0.0.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
